feat: added code optimization, splitted bill match code in diff components
